smbios: support UNIX-like operating systems, README intro
